\name{hgCLengths}
\alias{hgCLengths}
\non_function{}
\title{ A dataset which contains the lengths (in base pairs) of the
  human chromosomes. }
\usage{data(hgCLengths)}
\format{A vector containing 24 values, each corresponding to the total
chromosome length. }
\description{
  The data is described above.
}
\source{
  UCSC Human Genome Project
}
\examples{
  data(hgCLengths)
}
\keyword{datasets}
