\name{hgu95AProbLocs}
\docType{methods}
\alias{hgu95AProbLocs}
\title{chromLocation instance hgu95AProbLocs, an example of a chromLocation 
object}
\description{  gives chromosome locations for Affy U95 probes }
\section{Slots}{
  \describe{
    \item{\code{species}:}{Object of class character, value: 'Human'}
    \item{\code{datSource}:}{Object of class character, value}
    \item{\code{nChrom}:}{Object of class numeric, value: 24}
    \item{\code{chromNames}:}{Object of class character, value: 1:22, X,Y}
    \item{\code{chromLocs}:}{Object of class list, value: long: sense and antisense locations
      associated with affy identifiers}
    \item{\code{chromLengths}:}{Object of class numeric, }
    \item{\code{geneToChrom}:}{Object of class environment}
    \item{\code{class}:}{Object of class character, value: 'chromLocation'}
  }
}
\keyword{methods}
