\name{pm.getabst}
\alias{pm.getabst}
\title{ Obtain the abstracts for a set PubMed list. }
\description{
 The data provided by PubMed is reduced to a small set. This set is 
then suitable for further rendering.
}
\usage{
pm.getabst(affyids, basename)
}
\arguments{
  \item{affyids}{The identifiers used to find Abstracts }
  \item{basename}{The base name of the annotation package to use. }
}
\details{

}
\value{
  ~Describe the value returned
  If it is a LIST, use
  \item{comp1 }{Description of `comp1'}
  \item{comp2 }{Description of `comp2'}
  ...
}
\references{ ~put references to the literature/web site here ~ }
\author{ ~~who you are~~ }
\note{ ~~further notes~~ }

\seealso{ ~~objects to SEE ALSO as \code{\link{~~fun~~}}, ~~~ }

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
}
\keyword{ ~kwd1 }
