\name{annotate-defunct}
\alias{annotate-defunct}
\title{Defunct Functions in Package \pkg{annotate}}
%------ PLEASE: put \alias{.} here for  EACH !
\alias{neighborGeneFinder}
\alias{genelocator}
\alias{getQuery4LL}
%
\description{
  The functions or variables listed here are no longer part of the annotate package.
}
\usage{
neighborGeneFinder()
genelocator()
getQuery4LL()
}
\seealso{
  \code{\link{Defunct}}
}

